requirements.txt
docker
fastapi
uvicorn
websockets
pydantic[email]
python-dateutil
PyYAML
yagmail